backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
marshmallow>=3.20.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
gunicorn>=21.0.0
